+++ LDRSensorA5/Models/ManualModeData.cs
@@ -4,7 +4,7 @@ namespace LDRSensorA5.Models
 {
     public class ManualModeData
     {
-        public float Current { get; set; }  
+        public double Current { get; set; }  
         public bool RelayState { get; set; }
     }
 }
